requirements.txt
aws-cdk-lib==2.73.0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.73.0a0
aws-cdk.aws-apigatewayv2-authorizers-alpha==2.73.0a0
constructs>=10.0.0,<11.0.0
python-dotenv==1.0.0
